restructure __test__ folder
