+++ __tests__/unit/enemy.unit.test.js
@@ -2,10 +2,10 @@ const {
     ENUM_ENEMY_ATTACK, 
     ENUM_ENEMY_TYPE,
     ENUM_ENEMY_STRENGTH
-} = require('../generic/enums')
-const enemyAttackBuilder = require('../build/enemyAttack')
-const { getRandomNumberInRange } = require('../lib/utils')
-const b = require('../build/monster')
+} = require('../../generic/enums')
+const enemyAttackBuilder = require('../../build/enemyAttack')
+const { getRandomNumberInRange } = require('../../lib/utils')
+const b = require('../../build/monster')
 
 describe('enemy attack tests', () => {
     test('get hit attack', () =>  {
